renamed get app access token to facebook helpers
