+++ src/lib/facebook-helpers.js
